Fixed printf size_t formats
git-svn-id: https://gpsdrive.svn.sourceforge.net/svnroot/gpsdrive/trunk@1075 e3ca9d6f-f84a-410d-b97a-e7ddc331358d

+++ src/map_handler.c
@@ -304,7 +304,7 @@ make_display_map_checkboxes()
     }
 
   tooltips = gtk_tooltips_new ();
-  gint i;
+  glong i;
   for (i = 0; i < max_display_map; i++)
     {
       // Checkbox ---- Show Map: name xy
